refactor: create server/utils/ with shared email and parse utilities
Create canonical implementations in server/utils/:
- utils/email.js: normalizeEmail(), maskEmail()
- utils/parse.js: parseCsv(), parseDate(), parseZoomDate()

Eliminate duplicate code across the codebase:
- server/server.js: remove inline normalizeEmail/maskEmail (use utils/)
- server/services/spLedger.js: remove inline maskEmail (use utils/)
- server/services/sp.js: remove inline normalizeEmail/maskEmail (use utils/)
- server/scripts/lib/ingestion.js: re-export from utils, remove local duplicates

All call sites now import from the canonical source, so future bug fixes
and improvements only need to be made in one place.
